Map and Filter
The built-in functions map and filter are very useful higher-order functions that 
operate on lists (or similar objects called iterables). 
The function map takes a function and an iterable as arguments,
and returns a new iterable with the function applied to each argument.

Filter :- The function filter filters an iterable by removing items that don't match a predicate (a function that returns a Boolean).